requirements.txt
opencv-python
numpy
pillow
torch
torchvision
transformers